+++ src/tests/ascent/t_ascent_scalar_renderer.cpp
@@ -58,7 +58,7 @@ TEST(ascent_scalar_rendering, test_scalar_rendering)
 
     EXPECT_TRUE(conduit::blueprint::mesh::verify(data,verify_info));
 
-    ASCENT_INFO("Testing scalar  Rendering");
+    ASCENT_INFO("Testing Scalar Rendering");
 
 
     string output_path = prepare_output_dir();
@@ -98,7 +98,6 @@ TEST(ascent_scalar_rendering, test_scalar_rendering)
     //
 
     Ascent ascent;
-
     Node ascent_opts;
     ascent_opts["runtime/type"] = "ascent";
     ascent.open(ascent_opts);
@@ -108,6 +107,85 @@ TEST(ascent_scalar_rendering, test_scalar_rendering)
 
     // check that we created an image
     std::string msg = "An example of scalar rendering";
+    ASCENT_ACTIONS_DUMP(actions,output_file,msg);
+}
+
+
+//-----------------------------------------------------------------------------
+TEST(ascent_scalar_rendering, test_scalar_rendering_fields_specified)
+{
+    // the vtkm runtime is currently our only rendering runtime
+    Node n;
+    ascent::about(n);
+    // only run this test if ascent was built with vtkm support
+    if(n["runtimes/ascent/vtkm/status"].as_string() == "disabled")
+    {
+        ASCENT_INFO("Ascent support disabled, skipping test");
+        return;
+    }
+
+
+    //
+    // Create an example mesh.
+    //
+    Node data, verify_info;
+    conduit::blueprint::mesh::examples::braid("hexs",
+                                              EXAMPLE_MESH_SIDE_DIM,
+                                              EXAMPLE_MESH_SIDE_DIM,
+                                              EXAMPLE_MESH_SIDE_DIM,
+                                              data);
+
+    EXPECT_TRUE(conduit::blueprint::mesh::verify(data,verify_info));
+
+    ASCENT_INFO("Testing Scalar Rendering with fields specified");
+
+
+    string output_path = prepare_output_dir();
+    string output_file = conduit::utils::join_file_path(output_path,"tout_scalar_rendering_fields_specified");
+
+    //
+    // Create the actions.
+    //
+
+    conduit::Node pipelines;
+    // pipeline 1
+    pipelines["pl1/f1/type"] = "project_2d";
+    // filter knobs
+    conduit::Node &params = pipelines["pl1/f1/params"];
+    params["image_width"] = 512;
+    params["image_height"] = 512;
+    params["fields"].append() = "braid";
+
+    conduit::Node extracts;
+    extracts["e1/type"]  = "relay";
+    extracts["e1/pipeline"] = "pl1";
+
+    extracts["e1/params/path"] = output_file;
+    extracts["e1/params/protocol"] = "blueprint/mesh/hdf5";
+
+    conduit::Node actions;
+    // add the extracts
+    conduit::Node &add_extracts = actions.append();
+    add_extracts["action"] = "add_extracts";
+    add_extracts["extracts"] = extracts;
+    // add the pipeline
+    conduit::Node &add_pipelines= actions.append();
+    add_pipelines["action"] = "add_pipelines";
+    add_pipelines["pipelines"] = pipelines;
+
+    //
+    // Run Ascent
+    //
+
+    Ascent ascent;
+    Node ascent_opts;
+    ascent.open(ascent_opts);
+    ascent.publish(data);
+    ascent.execute(actions);
+    ascent.close();
+
+    // check that we created an image
+    std::string msg = "An example of scalar rendering of specific fields";
     ASCENT_ACTIONS_DUMP(actions,output_file,msg);
 }